+++ e2e/test_zsh.py
@@ -11,22 +11,11 @@ from conftest import set_working_dir
 from common import LSOF_FILTER_GRAMMAR, STRACE_EXPR_GRAMMAR
 
 
-def zsh_completions_from_stdout(stdout: str) -> list[tuple[str, str]]:
-    completions = []
-    for line in stdout.splitlines():
-        match = re.match(r'^(\S+)\s+--\s+(.+)$', line)
-        if match is not None:
-            completions.append((match.group(1), match.group(2)))
-        else:
-            completions.append((line, ''))
-    return completions
-
-
-def get_sorted_completions(generated_script_path: Path, input: str) -> list[tuple[str, str]]:
+def get_sorted_completions(generated_script_path: Path, input: str) -> list[str]:
     zsh_process = subprocess.run(['zsh', generated_script_path, input], stdout=subprocess.PIPE, stderr=sys.stderr, check=True)
     stdout = zsh_process.stdout.decode()
-    completions = zsh_completions_from_stdout(stdout)
-    completions.sort(key=lambda pair: pair[0])
+    completions = stdout.splitlines()
+    completions.sort()
     return completions
 
 
@@ -64,7 +53,7 @@ cmd <COMMAND> [--help];
 '''
 
     with capture_grammar_completions(complgen_binary_path, GRAMMAR) as capture_zsh_path:
-        assert get_sorted_completions(capture_zsh_path, 'cmd ') == sorted([('rm', "Remove a project"), ('remote', "Manage a project's remotes")], key=lambda pair: pair[0])
+        assert get_sorted_completions(capture_zsh_path, 'cmd ') == sorted(['rm rm     -- Remove a project', "remote remote -- Manage a project's remotes"])
 
 
 def test_zsh_uses_correct_description_with_duplicated_descriptions(complgen_binary_path: Path):
@@ -77,7 +66,10 @@ mygrep [<OPTION>]...;
 '''
 
     with capture_grammar_completions(complgen_binary_path, GRAMMAR) as capture_zsh_path:
-        assert get_sorted_completions(capture_zsh_path, 'mygrep ') == sorted([('--color', "use markers to highlight the matching strings"), ('--colour', "use markers to highlight the matching strings")], key=lambda pair: pair[0])
+        assert get_sorted_completions(capture_zsh_path, 'mygrep ') == sorted([
+            '--color --color  -- use markers to highlight the matching strings',
+            '--colour --colour -- use markers to highlight the matching strings',
+        ])
 
 
 def test_zsh_external_command_produces_description(complgen_binary_path: Path):
@@ -85,7 +77,7 @@ def test_zsh_external_command_produces_description(complgen_binary_path: Path):
 cmd { echo -e "completion\tdescription" };
 '''
     with capture_grammar_completions(complgen_binary_path, GRAMMAR) as capture_zsh_path:
-        assert get_sorted_completions(capture_zsh_path, 'cmd ') == sorted([('completion', 'description')])
+        assert get_sorted_completions(capture_zsh_path, 'cmd ') == sorted(['completion completion -- description'])
 
 
 def test_completes_paths(complgen_binary_path: Path):
@@ -95,7 +87,11 @@ def test_completes_paths(complgen_binary_path: Path):
                 Path('filename with spaces').write_text('dummy')
                 Path('?[^a]*{foo,*bar}').write_text('dummy')
                 os.mkdir('dir with spaces')
-                assert get_sorted_completions(capture_zsh_path, 'cmd ') == sorted([(r'\?\[\^a\]\*\{foo,\*bar\}', ''), (r'filename\ with\ spaces', ''), (r'dir\ with\ spaces', '')])
+                assert get_sorted_completions(capture_zsh_path, 'cmd ') == sorted([
+                    r'\?\[\^a\]\*\{foo,\*bar\}',
+                    r'filename\ with\ spaces',
+                    r'dir\ with\ spaces',
+                ])
 
 
 def test_completes_directories(complgen_binary_path: Path):
@@ -105,7 +101,10 @@ def test_completes_directories(complgen_binary_path: Path):
                 os.mkdir('dir with spaces')
                 os.mkdir('?[^a]*{foo,*bar}')
                 Path('filename with spaces').write_text('dummy')
-                assert get_sorted_completions(capture_zsh_path, 'cmd ') == sorted([(r'\?\[\^a\]\*\{foo,\*bar\}', ''), (r'dir\ with\ spaces', '')])
+                assert get_sorted_completions(capture_zsh_path, 'cmd ') == sorted([
+                    r'\?\[\^a\]\*\{foo,\*bar\}',
+                    r'dir\ with\ spaces',
+                ])
 
 
 def test_completes_file_with_spaces(complgen_binary_path: Path):
@@ -113,7 +112,7 @@ def test_completes_file_with_spaces(complgen_binary_path: Path):
         with tempfile.TemporaryDirectory() as dir:
             with set_working_dir(Path(dir)):
                 Path('file with spaces').write_text('dummy')
-                assert get_sorted_completions(capture_zsh_path, 'cmd ') == sorted([('file\\ with\\ spaces', '')])
+                assert get_sorted_completions(capture_zsh_path, 'cmd ') == sorted(['file\\ with\\ spaces'])
 
 
 def get_jit_zsh_completions_expr(complgen_binary_path: Path, grammar: str, completed_word_index: int, words_before_cursor: list[str]) -> str:
@@ -131,7 +130,7 @@ def test_jit_completes_paths_zsh(complgen_binary_path: Path):
             assert expr.splitlines() == [
                 r'local -a completions=("\?\[\^a\]\*\{foo,\*bar\}" "dir\ with\ spaces" "filename\ with\ spaces")',
                 r'local -a descriptions=("\?\[\^a\]\*\{foo,\*bar\}" "dir\ with\ spaces" "filename\ with\ spaces")',
-                '''compadd -Q -S '' -d descriptions -a completions'''
+                '''compadd -l -Q -S '' -d descriptions -a completions'''
             ]
 
 
@@ -145,7 +144,7 @@ def test_jit_completes_directories_zsh(complgen_binary_path: Path):
             assert expr.splitlines() == [
                 r'local -a completions=("\?\[\^a\]\*\{foo,\*bar\}" "dir\ with\ spaces")',
                 r'local -a descriptions=("\?\[\^a\]\*\{foo,\*bar\}" "dir\ with\ spaces")',
-                '''compadd -Q -S '' -d descriptions -a completions'''
+                '''compadd -l -Q -S '' -d descriptions -a completions'''
             ]
 
 
@@ -158,19 +157,19 @@ def test_jit_completes_subdirectory_files(complgen_binary_path: Path):
             assert expr.splitlines() == [
                 r'local -a completions=("subdir/file.txt")',
                 r'local -a descriptions=("subdir/file.txt")',
-                '''compadd -Q -S '' -d descriptions -a completions'''
+                '''compadd -l -Q -S '' -d descriptions -a completions'''
             ]
 
 
 
 def test_specializes_for_zsh(complgen_binary_path: Path):
     with capture_grammar_completions(complgen_binary_path, '''cmd <FOO>; <FOO> ::= { echo foo }; <FOO@zsh> ::= { compadd zsh };''') as capture_zsh_path:
-        assert get_sorted_completions(capture_zsh_path, 'cmd ') == sorted([('zsh', '')])
+        assert get_sorted_completions(capture_zsh_path, 'cmd ') == sorted(['zsh'])
 
 
 def test_jit_specializes_for_zsh(complgen_binary_path: Path):
     expr = get_jit_zsh_completions_expr(complgen_binary_path, '''cmd <FOO>; <FOO> ::= { echo foo }; <FOO@zsh> ::= { compadd zsh };''', 0, [])
-    assert expr == '''local -a completions=("zsh")\nlocal -a descriptions=("zsh")\ncompadd -Q -S '' -d descriptions -a completions\n'''
+    assert expr == '''local -a completions=("zsh")\nlocal -a descriptions=("zsh")\ncompadd -l -Q -S '' -d descriptions -a completions\n'''
 
 
 def test_mycargo(complgen_binary_path: Path):
@@ -182,7 +181,7 @@ cargo [<toolchain>] [<COMMAND>];
 '''
 
     with capture_grammar_completions(complgen_binary_path, GRAMMAR) as capture_zsh_path:
-        assert get_sorted_completions(capture_zsh_path, 'cargo t ') == sorted([('testname', '')])
+        assert get_sorted_completions(capture_zsh_path, 'cargo t ') == sorted(['testname testname'])
 
 
 def test_matches_prefix(complgen_binary_path: Path):
@@ -192,7 +191,7 @@ cargo test --test testname;
 <toolchain> ::= stable-aarch64-apple-darwin | stable-x86_64-apple-darwin;
 '''
     with capture_grammar_completions(complgen_binary_path, GRAMMAR) as capture_zsh_path:
-        assert get_sorted_completions(capture_zsh_path, 'cargo +stable-aarch64-apple-darwin ') == sorted([('foo', '')])
+        assert get_sorted_completions(capture_zsh_path, 'cargo +stable-aarch64-apple-darwin ') == sorted(['foo foo'])
 
 
 def test_jit_matches_prefix(complgen_binary_path: Path):
@@ -201,7 +200,7 @@ cargo +<toolchain> foo;
 <toolchain> ::= stable-aarch64-apple-darwin | stable-x86_64-apple-darwin;
 '''
     expr = get_jit_zsh_completions_expr(complgen_binary_path, GRAMMAR, 1, ['+stable-aarch64-apple-darwin'])
-    assert expr == '''local -a completions=("foo")\nlocal -a descriptions=("foo")\ncompadd -Q -S '' -d descriptions -a completions\n'''
+    assert expr == '''local -a completions=("foo")\nlocal -a descriptions=("foo")\ncompadd -l -Q -S '' -d descriptions -a completions\n'''
 
 
 def test_completes_prefix(complgen_binary_path: Path):
@@ -210,7 +209,7 @@ cargo +<toolchain>;
 <toolchain> ::= stable-aarch64-apple-darwin | stable-x86_64-apple-darwin;
 '''
     with capture_grammar_completions(complgen_binary_path, GRAMMAR) as capture_zsh_path:
-        assert get_sorted_completions(capture_zsh_path, 'cargo +') == sorted([('+stable-aarch64-apple-darwin', ''), ('+stable-x86_64-apple-darwin', '')])
+        assert get_sorted_completions(capture_zsh_path, 'cargo +') == sorted(['+stable-aarch64-apple-darwin +stable-aarch64-apple-darwin', '+stable-x86_64-apple-darwin +stable-x86_64-apple-darwin'])
 
 
 def test_jit_completes_prefix(complgen_binary_path: Path):
@@ -219,50 +218,50 @@ cargo +<toolchain>;
 <toolchain> ::= stable-aarch64-apple-darwin | stable-x86_64-apple-darwin;
 '''
     expr = get_jit_zsh_completions_expr(complgen_binary_path, GRAMMAR, 0, ['+'])
-    assert expr == '''local -a completions=("+stable-aarch64-apple-darwin" "+stable-x86_64-apple-darwin")\nlocal -a descriptions=("+stable-aarch64-apple-darwin" "+stable-x86_64-apple-darwin")\ncompadd -Q -S '' -d descriptions -a completions\n'''
+    assert expr == '''local -a completions=("+stable-aarch64-apple-darwin" "+stable-x86_64-apple-darwin")\nlocal -a descriptions=("stable-aarch64-apple-darwin" "stable-x86_64-apple-darwin")\ncompadd -l -Q -S '' -d descriptions -a completions\n'''
 
 
 def test_completes_strace_expr(complgen_binary_path: Path):
     with capture_grammar_completions(complgen_binary_path, STRACE_EXPR_GRAMMAR) as capture_zsh_path:
-        assert get_sorted_completions(capture_zsh_path, 'strace -e ') == sorted([('%file', ''), ('!', ''), ('all', ''), ('fault', ''), ('file', ''), ('read', ''), ('trace', ''), ('write', '')])
+        assert get_sorted_completions(capture_zsh_path, 'strace -e ') == sorted(['%file %file', '! !', 'all all', 'fault fault', 'file file', 'read read', 'trace trace', 'write write'])
 
 
 def test_jit_completes_strace_expr(complgen_binary_path: Path):
     expr = get_jit_zsh_completions_expr(complgen_binary_path, STRACE_EXPR_GRAMMAR, 1, ['-e', 'trace='])
-    assert expr == '''local -a completions=("trace=!" "trace=%file" "trace=all" "trace=file")\nlocal -a descriptions=("trace=!" "trace=%file" "trace=all" "trace=file")\ncompadd -Q -S '' -d descriptions -a completions\n'''
+    assert expr == '''local -a completions=("trace=!" "trace=%file" "trace=all" "trace=file")\nlocal -a descriptions=("!" "%file" "all" "file")\ncompadd -l -Q -S '' -d descriptions -a completions\n'''
 
 
 def test_completes_lsof_filter(complgen_binary_path: Path):
     with capture_grammar_completions(complgen_binary_path, LSOF_FILTER_GRAMMAR) as capture_zsh_path:
-        assert get_sorted_completions(capture_zsh_path, 'lsf -sTCP:') == sorted([('-sTCP:LISTEN', ''), ('-sTCP:CLOSED', ''), ('-sTCP:^', '')])
+        assert get_sorted_completions(capture_zsh_path, 'lsf -sTCP:') == sorted(['-sTCP:LISTEN -sTCP:LISTEN', '-sTCP:CLOSED -sTCP:CLOSED', '-sTCP:^ -sTCP:^'])
 
 
 def test_jit_completes_lsof_filter(complgen_binary_path: Path):
     expr = get_jit_zsh_completions_expr(complgen_binary_path, LSOF_FILTER_GRAMMAR, 0, ['-sTCP:'])
-    assert expr == '''local -a completions=("-sTCP:CLOSED" "-sTCP:LISTEN" "-sTCP:^")\nlocal -a descriptions=("-sTCP:CLOSED" "-sTCP:LISTEN" "-sTCP:^")\ncompadd -Q -S '' -d descriptions -a completions\n'''
+    assert expr == '''local -a completions=("-sTCP:CLOSED" "-sTCP:LISTEN" "-sTCP:^")\nlocal -a descriptions=("CLOSED" "LISTEN" "^")\ncompadd -l -Q -S '' -d descriptions -a completions\n'''
 
 
 def test_subword_descriptions(complgen_binary_path: Path):
     GRAMMAR = r'''cmd --option=(arg1 "descr1" | arg2 "descr2");'''
     with capture_grammar_completions(complgen_binary_path, GRAMMAR) as capture_zsh_path:
-        assert get_sorted_completions(capture_zsh_path, 'cmd --option=') == sorted([('--option=arg1', 'arg1 (descr1)'), ('--option=arg2', 'arg2 (descr2)')])
+        assert get_sorted_completions(capture_zsh_path, 'cmd --option=') == sorted(['--option=arg1 --option=arg1 -- arg1 (descr1)', '--option=arg2 --option=arg2 -- arg2 (descr2)'])
 
 def test_jit_subword_descriptions(complgen_binary_path: Path):
     GRAMMAR = r'''cmd --option=(arg1 "descr1" | arg2 "descr2");'''
     expr = get_jit_zsh_completions_expr(complgen_binary_path, GRAMMAR, 0, ['--option='])
-    assert expr == '''local -a completions=("--option=arg1" "--option=arg2")\nlocal -a descriptions=("--option=arg1 -- descr1" "--option=arg2 -- descr2")\ncompadd -Q -S '' -d descriptions -a completions\n'''
+    assert expr == '''local -a completions=("--option=arg1" "--option=arg2")\nlocal -a descriptions=("arg1 -- descr1" "arg2 -- descr2")\ncompadd -l -Q -S '' -d descriptions -a completions\n'''
 
 
 def test_completes_subword_external_command(complgen_binary_path: Path):
     GRAMMAR = r'''cmd --option={ echo -e "argument\tdescription" };'''
     with capture_grammar_completions(complgen_binary_path, GRAMMAR) as capture_zsh_path:
-        assert get_sorted_completions(capture_zsh_path, 'cmd --option=') == sorted([('--option=argument', 'description')])
+        assert get_sorted_completions(capture_zsh_path, 'cmd --option=') == sorted(['--option=argument --option=argument -- description'])
 
 
 def test_jit_completes_subword_external_command(complgen_binary_path: Path):
     GRAMMAR = r'''cmd --option={ echo -e "argument\tdescription" };'''
     expr = get_jit_zsh_completions_expr(complgen_binary_path, GRAMMAR, 0, ['--option='])
-    assert expr == '''local -a completions=("--option=argument")\nlocal -a descriptions=("--option=argument -- description")\ncompadd -Q -S '' -d descriptions -a completions\n'''
+    assert expr == '''local -a completions=("--option=argument")\nlocal -a descriptions=("argument -- description")\ncompadd -l -Q -S '' -d descriptions -a completions\n'''
 
 
 def test_subword_specialization(complgen_binary_path: Path):
@@ -272,7 +271,7 @@ cmd --option=<FOO>;
 <FOO@zsh> ::= { echo zsh };
 '''
     with capture_grammar_completions(complgen_binary_path, GRAMMAR) as capture_zsh_path:
-        assert get_sorted_completions(capture_zsh_path, 'cmd --option=') == sorted([('--option=zsh', '')])
+        assert get_sorted_completions(capture_zsh_path, 'cmd --option=') == sorted(['--option=zsh --option=zsh'])
 
 
 def test_jit_subword_specialization(complgen_binary_path: Path):
@@ -282,7 +281,7 @@ cmd --option=<FOO>;
 <FOO@zsh> ::= { echo zsh };
 '''
     expr = get_jit_zsh_completions_expr(complgen_binary_path, GRAMMAR, 0, ['--option='])
-    assert expr == '''local -a completions=("zsh")\nlocal -a descriptions=("zsh")\ncompadd -Q -S '' -d descriptions -a completions\n'''
+    assert expr == '''local -a completions=("zsh")\nlocal -a descriptions=("zsh")\ncompadd -l -Q -S '' -d descriptions -a completions\n'''
 
 
 def test_shell_integration(complgen_binary_path: Path):
@@ -309,6 +308,6 @@ done
         with capture_script_path(INTEGRATION_SCRIPT) as capture_path:
             zsh_process = subprocess.run(['zsh', capture_path, 'mycargo +'], stdout=subprocess.PIPE, stderr=sys.stderr, check=True)
             stdout = zsh_process.stdout.decode()
-            completions = zsh_completions_from_stdout(stdout)
-            completions.sort(key=lambda pair: pair[0])
-            assert completions == [('+bar', ''), ('+foo', '')]
+            completions = stdout.splitlines()
+            completions.sort()
+            assert completions == ['+bar bar', '+foo foo']